+++ config/splat.us.stnp3.yaml
@@ -49,11 +49,14 @@ segments:
       - [0x3A7C, roomdef, g_Rooms]
       - [0x3B68, rodata]
       - [0x31EA0, rodata]
+      - [0x31F88, .rodata, 3E908]
+      - [0x31FA0, rodata]
       - [0x31FB0, rodata]
       - [0x3208C, .rodata, 4997C]
       - [0x320A0, rodata]
       - [0x32360, rodata]
       - [0x3246C, c]
+      - [0x3E908, c, 3E908]
       - [0x4997C, c, 4997C]
       - [0x52768, data]
   - [0x53434]
